Fix DM getting marked unread from your own message
